+++ R/bundle.R
@@ -676,10 +676,11 @@ createAppManifest <- function(appDir, appMode, contentCategory, hasParameters,
 
   # provide package entries for all dependencies
   packages <- list()
-  # non-SCM repository sources without URLs
-  missing_url_sources <- NULL
-  # potential error messages
-  msg      <- NULL
+
+  # potential problems with the bundled content.
+  errorMessages <- NULL
+  packageMessages <- NULL
+
   pyInfo   <- NULL
 
   # get package dependencies for non-static content deployment
@@ -702,21 +703,12 @@ createAppManifest <- function(appDir, appMode, contentCategory, hasParameters,
           pyInfo$package_manager$contents <- NULL
         }
         else {
-          msg <- c(msg, paste("Error detecting python for reticulate:", pyInfo$error))
+          errorMessages <- c(errorMessages, paste("Error detecting python for reticulate:", pyInfo$error))
         }
       }
 
       # get package info
-      info <- as.list(deps[i, c('Source',
-                                'Repository')])
-
-      if (is.na(info$Repository)) {
-        if (isSCMSource(info$Source)) {
-          # ignore source+SCM packages
-        } else {
-          missing_url_sources <- unique(c(missing_url_sources, info$Source))
-        }
-      }
+      info <- as.list(deps[i, c('Source', 'Repository')])
 
       # include github package info
       info <- c(info, as.list(deps[i, grep('Github', colnames(deps), perl = TRUE, value = TRUE)]))
@@ -729,30 +721,37 @@ createAppManifest <- function(appDir, appMode, contentCategory, hasParameters,
 
       # if description is NA, application dependency may not be installed
       if (is.na(info$description[1])) {
-        msg <- c(msg, paste0(capitalize(assetTypeName), " depends on package \"",
+        errorMessages <- c(errorMessages, paste0(capitalize(assetTypeName), " depends on package \"",
                              name, "\" but it is not installed. Please resolve ",
                              "before continuing."))
         next
       }
 
-      # validate package source (returns an error message if there is a problem)
-      msg <- c(msg, validatePackageSource(deps[i, ]))
+      # validate package source (returns a message if there is a problem)
+      packageMessages <- c(packageMessages, validatePackageSource(deps[i, ]))
 
       # good to go
       packages[[name]] <- info
     }
   }
-  if (length(missing_url_sources)) {
-    # Err when packages lack repository URL. We emit a warning about each package (see
-    # snapshotDependencies) before issuing an error with this resolution advice.
+  if (length(packageMessages)) {
+    # Advice to help resolve installed packages that are not available using the current set of
+    # configured repositories. Each package with a missing repository has already been printed (see
+    # snapshotDependencies).
     #
-    # It's possible we cannot find a repository URL for other reasons, including when folks locally
-    # build and install packages from source. An incorrectly configured "repos" option is almost
-    # always the cause.
-    msg <- c(msg, sprintf("Unable to determine the location for some packages. Packages must come from a package repository like CRAN or a source control system. Check that options('repos') refers to a package repository containing the needed package versions."))
+    # This situation used to trigger an error (halting deployment), but was softened because:
+    #   * CRAN-archived packages are not visible to our available.packages scanning.
+    #   * Source-installed packages may be available after a manual server-side installation.
+    #
+    # That said, an incorrectly configured "repos" option is almost always the cause.
+    packageMessages <- c(packageMessages,
+                         "Unable to determine the location for some packages. Packages should be installed from a package repository like CRAN or a source control system. Check that options('repos') refers to a package repository containing the needed package versions.")
+    warning(paste(formatUL(packageMessages, '\n*'), collapse = '\n'), call. = FALSE, immediate. = TRUE)
   }
 
-  if (length(msg)) stop(paste(formatUL(msg, '\n*'), collapse = '\n'), call. = FALSE)
+  if (length(errorMessages)) {
+    stop(paste(formatUL(errorMessages, '\n*'), collapse = '\n'), call. = FALSE)
+  }
 
   if (!retainPackratDirectory) {
     # Optionally remove the packrat directory when it will not be included in
@@ -833,17 +832,16 @@ createAppManifest <- function(appDir, appMode, contentCategory, hasParameters,
 }
 
 validatePackageSource <- function(pkg) {
-  msg <- NULL
-  if (!(pkg$Source %in% c("CRAN", "Bioconductor", "github", "gitlab", "bitbucket"))) {
-    if (is.null(pkg$Repository)) {
-      msg <- paste("The package was installed from an unsupported ",
-                   "source '", pkg$Source, "'.", sep = "")
-    }
+  if (isSCMSource(pkg$Source)) {
+    return()
   }
-  if (is.null(msg)) return()
-  msg <- paste("Unable to deploy package dependency '", pkg$Package,
-               "'\n\n", msg, " ", sep = "")
-  msg
+
+  if (is.null(pkg$Repository) || is.na(pkg$Repository)) {
+    return(sprintf("May be unable to deploy package dependency '%s'; could not determine a repository URL for the source '%s'.",
+                   pkg$Package, pkg$Source))
+  }
+
+  return()
 }
 
 hasRequiredDevtools <- function() {