+++ NEWS.md
@@ -7,6 +7,10 @@
 * Added linter for different-case Markdown links (#388)
 * Use new Packrat release on CRAN, 0.6.0 (#501)
 * Fix incorrect linter messages referring to `shiny.R` instead of `server.R` (#509)
+* Warn, rather than err, when the repository URL for a package dependency
+  cannot be validated. This allows deployment when using archived CRAN
+  packages, or when using packages installed from source that are available on
+  the server. (#508)
 
 ## 0.8.17